+++ lib/core/router/app_route_manager.dart
@@ -1,26 +1,16 @@
-import 'package:clean_architecture_example/core/router/app_route_enum.dart';
-import 'package:clean_architecture_example/features/details/details_screen.dart';
-import 'package:clean_architecture_example/features/login/presentation/screen/login_screen.dart';
-import 'package:clean_architecture_example/features/splash/splash_screen.dart';
+import 'package:clean_architecture_example/features/details/core/router/router.dart';
+import 'package:clean_architecture_example/features/login/core/router/router.dart';
+import 'package:clean_architecture_example/features/splash/core.router/router.dart';
 import 'package:go_router/go_router.dart';
 
 class AppRouteManager {
   AppRouteManager._();
   static final GoRouter router = GoRouter(
-    initialLocation: AppRoute.splash.fullPath,
+    initialLocation: SplashRouter.routePath,
     routes: [
-      GoRoute(
-        path: AppRoute.splash.fullPath,
-        builder: (context, state) => SplashScreen(),
-      ),
-      GoRoute(
-        path: AppRoute.login.fullPath,
-        builder: (context, state) => LoginScreen(),
-      ),
-      GoRoute(
-        path: AppRoute.details.fullPath,
-        builder: (context, state) => DetailsScreen(),
-      ),
+      SplashRouter.splashRouter,
+      LoginRouter.loginRouter,
+      UserDetailsRouter.userDetailsRouter,
     ],
   );
 }