+++ lib/features/details/data/params/user_details_params.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'user_details_params.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$UserDetailsParamsImpl _$$UserDetailsParamsImplFromJson(
+        Map<String, dynamic> json) =>
+    _$UserDetailsParamsImpl(
+      token: json['token'] as String?,
+    );
+
+Map<String, dynamic> _$$UserDetailsParamsImplToJson(
+        _$UserDetailsParamsImpl instance) =>
+    <String, dynamic>{
+      'token': instance.token,
+    };