+++ lib/features/details/presentation/bloc/api/user_details_event.dart
@@ -1,0 +1,10 @@
+part of 'user_details_bloc.dart';
+
+@freezed
+class UserDetailsEvent with _$UserDetailsEvent {
+  const factory UserDetailsEvent.updateResponse({
+    required UserDetailsParams params,
+  }) = _$UpdateResponse;
+
+  const factory UserDetailsEvent.resetBloc() = _ResetBloc;
+}